+++ src/JS.hs
@@ -1,0 +1,11 @@
+module JS
+    ( module JS.Types
+    , module JS.Cast
+    , module JS.Data
+    ) where
+
+import JS.Aeson.Instances ()
+import JS.Cast
+import JS.Data
+import JS.Data.Instances ()
+import JS.Types