+++ src/JS/Data.hs
@@ -1,0 +1,10 @@
+module JS.Data
+    ( JSVal
+    , nullRef
+    , JSString
+    , module JS.Data.Object
+    ) where
+
+import GHCJS.Types
+import JS.Data.Instances ()
+import JS.Data.Object