+++ src/JS/Data/Unsafe.hs
@@ -1,13 +1,13 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 {-# LANGUAGE CPP #-}
 
-module JavaScript.Extras.JSVal.Unsafe where
+module JS.Data.Unsafe where
 
 import Data.JSString as J
-import qualified GHCJS.Types as J
+import GHCJS.Types
 
 -- | Injection attack! Use with care
-instance Read J.JSVal where
+instance Read JSVal where
     readsPrec _ str = [(js_eval (J.pack str), [])]
 
 #ifdef __GHCJS__
@@ -15,11 +15,11 @@ instance Read J.JSVal where
 -- | Injection attack! Use with care
 foreign import javascript unsafe
   "$r = hje$unstringify($1);"
-  js_eval :: J.JSString -> J.JSVal
+  js_eval :: JSString -> JSVal
 
 #else
 
-js_eval :: J.JSString -> J.JSVal
-js_eval _ = J.nullRef
+js_eval :: JSString -> JSVal
+js_eval _ = nullRef
 
 #endif