+++ src/JS/Types.hs
@@ -1,0 +1,13 @@
+module JS.Types
+    ( JSVal
+    , JSString
+    , JSObject
+    , Callback
+    , Export
+    ) where
+
+import Data.JSString
+import GHCJS.Foreign.Callback
+import GHCJS.Foreign.Export
+import GHCJS.Types
+import JS.Data.Object