+++ src/JavaScript/Extras/JSString.hs
@@ -1,7 +1,0 @@
-module JavaScript.Extras.JSString where
-
-import qualified Data.JSString as J
-
--- | Convenience function when using OverloadedStrings
-jsstr :: J.JSString -> J.JSString
-jsstr = id